+++ api/src/main/java/com/exelaration/abstractmemery/services/ImageService.java
@@ -1,36 +1,16 @@
 package com.exelaration.abstractmemery.services;
 
-import java.util.List;
-import java.util.Optional;
-
 import javax.transaction.Transactional;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.exelaration.abstractmemery.domains.Image;
-import com.exelaration.abstractmemery.repositories.MemeRepository;
 
 @Service
 @Transactional
-public class ImageService {
+public interface ImageService {
 
-    @Autowired 
-    private MemeRepository memeRepository;
+    public Image save(MultipartFile file);
 
-    public List<Image> findAll() {
-        return (List<Image>) memeRepository.findAll();
-    }
-
-    public Optional<Image> findById(Long id) {
-        return memeRepository.findById(id);
-    }
-
-    public Image save(Image image) {
-        return memeRepository.save(image);
-    }
-
-    public void deleteById(Long id) {
-        memeRepository.deleteById(id);
-    }
 }
